feat(lsp): move to native lsp config
