+++ init.lua
@@ -1,8 +1,10 @@
+require 'core.mason-path'
 require 'config.options'
 require 'config.keymaps'
 require 'config.autocommands'
 require 'config.usercommands'
 
 require 'core.lazy'
+require 'core.lsp'
 
 require 'config.lazyblock_info'
