+++ lua/plugins/fidget.lua
@@ -1,0 +1,5 @@
+return {
+  'j-hui/fidget.nvim',
+  event = 'LspAttach',
+  opts = {},
+}
