+++ lua/plugins/mason.lua
@@ -1,0 +1,4 @@
+return {
+  'mason-org/mason.nvim',
+  opts = {},
+}
